dotenv path configured and refactored code into different archives
