+++ .env
@@ -1,2 +1,2 @@
-DATABASE_URL=postgressql://testando123:senha123@localhost5432/database123
+DATABASE_URL=postgres://testando123:senha123@localhost:5432/database123
 PORT=4000
