+++ package.json
@@ -31,7 +31,8 @@
       "build": "tsc",
       "start": "tsc & node build/index.js",
       "dev-start": "ts-node-dev src/index.ts",
-      "test": "echo \"Error: no test specified\" && exit 1"
+      "test-dev": "env TEST=true ts-node-dev src/index.ts"
+
    },
    "repository": {
       "type": "git",
