+++ src/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './resolvers';
+export * from './typeDefs';
