+++ test.env
@@ -1,2 +1,2 @@
-DATABASE_URL=postgressql://testando123:senha123@localhost5433/database123
+DATABASE_URL=postgres://testando123:senha123@localhost:5433/database123
 PORT=4001
